+++ jaguar/lib/http/mux/group_builder.dart
@@ -15,9 +15,9 @@ class GroupBuilder extends Object with Muxable {
 
   Route addRoute(Route route) {
     final route1 = route.cloneWith(path: pathPrefix + route.info.path);
-    route1.before.insertAll(0, before);
-    route1.after.insertAll(0, after);
-    route1.onException.insertAll(0, onException);
+    route1._before.insertAll(0, before);
+    route1._after.insertAll(0, after);
+    route1._onException.insertAll(0, onException);
     server.addRoute(route1);
     return route1;
   }